+++ pom.xml
@@ -143,12 +143,6 @@
                 <configuration>
                     <!--needed to prevent module problem-->
                     <detectJavaApiLink>false</detectJavaApiLink>
-                    <sourcepath>
-                        ${basedir}/src/main/java/calculator;
-                        ${basedir}/src/main/java/visitor;
-                        ${basedir}/src/main/java/gui;
-                        ${basedir}/src/main/java/common
-                    </sourcepath>
                 </configuration>
                 <executions>
                     <!-- Exports JavaDocs to regular HTML files -->
