--- src/main/java/module-info.java
@@ -1,9 +1,0 @@
-module gui.calculator {
-    requires javafx.controls;
-    requires javafx.fxml;
-    requires org.kordamp.bootstrapfx.core;
-    requires org.antlr.antlr4.runtime;
-
-    opens gui to javafx.fxml;
-    exports gui;
-}
